refactor(dsv4): relocate cache + sparse_index helpers under ops/attention/
Two DSV4-specific modules were sitting loose at ops/ root level alongside
generic primitives (_custom_ops / native_ops / deep_gemm etc), inconsistent
with baseline's "organize by functional role" convention. Relocated both
into existing ``ops/attention/`` subdirectory nextdoor to their natural
semantic siblings already housed there -- compressor.py / indexer_decode.py /
norms.py / flashmla_bridge.py / qkv_cache_insert.py / o_proj_alias.py --
which collectively form the MLA-decode support layer these two belong with:

  vllm_kunlun/ops/deepseek_v4_cache.py        -> vllm_kunlun/ops/attention/deepseek_v4_cache.py
  vllm_kunlun/ops/deepseek_v4_sparse_index.py -> vllm_kunlun/ops/attention/deepseek_v4_sparse_index.py

Basenames preserved verbatim so git rename detection tracks moves cleanly.
Consumer footprint was minimal: only THREE single-line import-path string
substitutions across exactly TWO consumer files
(tests/ut/test_deepseek_v4_cache.py L4; patches/dsv4_cache.py L16+L19),
NO behavioral changes, NO predicate-string identity-comparison updates needed,
no other residual references anywhere in repo per pre-execution grep audit.

Remaining candidates involving __module__ coupling stay deferred pending
separate evaluation honoring user's preference for minimizing edits to
existing code so as not to introduce rebase-conflict surface against upstream.

Verification:

* py_compile clean for every touched Python file.
* whitespace check returns OK.
* Focused UT unchanged vs HEAD~0 baseline: 21 passed / 2 pre-existing reds orthogonal.
* Live server PID 1949466 brought up cleanly post kill.sh cycle:
    /health=200 reached around t+150s cold-start mark,
    /v1/models returned DeepSeek-V4-Flash-INT8 JSON payload correctly,
    IrOpPriorityConfig(rms_norm='native', fused_add_rms_norm='native') milestone still applies twice end-to-end,
    Application startup complete logged normally.
* Fresh serve-log audit confirms no new non-deep_gemm-baseline ERROR / Traceback introduced;
  neither moved module emits explicit INFO on patch application so successful engine init +
  cudagraph capture + warmup completion serves as indirect proof substituted functions remained
  callable throughout startup sequence.
